+++ packages/pie-models/scripts/GenerateSchemas.ts
@@ -4,15 +4,6 @@ import { lstatSync, readdirSync, existsSync, writeFileSync, mkdirpSync, removeSy
 import {pascalCase, paramCase, camelCase} from 'change-case';
 
 
-interface PiePackageModelDefinition {
-  configModelFile: string;
-  itemModelFile: string;
-  packageName: string;
-}
-
-
-const pieDefinitions:Array<PiePackageModelDefinition> = [];
-
 const PIE_DEFINITIONS_DIR = "src/pie";
 const OUT_DIR = join('dist', 'schemas');
 
@@ -27,50 +18,48 @@ mkdirpSync(OUT_DIR);
 const getDirectories = (p: any) => readdirSync(p).filter(f => lstatSync(join(p, f)).isDirectory())
 const dirs = getDirectories(PIE_DEFINITIONS_DIR);
 
-dirs.forEach(dir => {
-  const itemModelFile = resolve(PIE_DEFINITIONS_DIR, dir, 'Pie.ts');
-  const configModelFile = resolve(PIE_DEFINITIONS_DIR, dir, 'Configure.ts');
-  if (existsSync(itemModelFile))  {
-    pieDefinitions.push({
-      packageName: dir,
-      itemModelFile,
-      configModelFile
-    });
-  }
-});
 
 
-
-pieDefinitions.forEach(def => {
-  const typePrefix = pascalCase(def.packageName);
-  const packageOutDir = join(OUT_DIR, def.packageName);
+const writeSchemaForPie = (packageName: string, filePath:string)  => {
+  const typePrefix = pascalCase(packageName);
+  const packageOutDir = join(OUT_DIR, packageName);
   mkdirpSync(packageOutDir);
 
   console.log(`generated ${packageOutDir}`);
 
   // load schemas
-  const files = [def.configModelFile, def.itemModelFile].filter(existsSync);
-  const program = TJS.getProgramFromFiles(files);
+  const program = TJS.getProgramFromFiles([filePath]);
   const generator = TJS.buildGenerator(program, SETTINGS);
 
   if (generator) {
    
     // write config model
-    if (files.some(x => x === def.configModelFile)) {
-      
+    try {
       const configSchema = generator.getSchemaForSymbol(typePrefix +  'Configure');
       writeFileSync(join(packageOutDir, 'config-schema.json'), JSON.stringify(configSchema, null, 2));
+    } catch (error) {
+      console.log(`no config available for ${packageName}`);
     }
-     // write item model
-    if (files.some(x => x === def.itemModelFile)) {
-      const itemSchema = generator.getSchemaForSymbol(typePrefix +  'Pie');
-      writeFileSync(join(packageOutDir, 'pie-schema.json'), JSON.stringify(itemSchema, null, 2));
+    try {
+      const pieSchema = generator.getSchemaForSymbol(typePrefix +  'Pie');
+      writeFileSync(join(packageOutDir, 'pie-schema.json'), JSON.stringify(pieSchema, null, 2));
+    } catch (error) {
+      console.log(`no pie model available for ${packageName}`);
     }
  
   } else {
-    console.error(`could not load schema files for ${def.packageName}`);
+    console.error(`could not load schema files for ${packageName}`);
   }
-  
+
+}
+
+
+dirs.forEach(dir => {
+  const modelFile = resolve(PIE_DEFINITIONS_DIR, dir, 'index.ts');
+  if (existsSync(modelFile))  {
+
+    writeSchemaForPie(dir, modelFile);
+  }
 });
