+++ packages/pie-models/src/index.ts
@@ -1,5 +1,4 @@
-export {MultipleChoicePie}  from './pie/multiple-choice/Pie';
-export {MultipleChoiceConfigure} from './pie/multiple-choice/Configure';
-export {CategorizePie} from './pie/categorize/Pie';
-export {CategorizeConfigure} from './pie/categorize/Configure';
+export {MultipleChoicePie, MultipleChoiceConfigure}  from './pie/multiple-choice';
+export {CategorizePie, CategorizeConfigure} from './pie/categorize';
+export {CalculatorPie} from './pie/calculator';
