+++ packages/pie-models/src/pie/multiple-choice/index.ts
@@ -1,7 +1,41 @@
+import {Choice}  from '../../Choice';
+
+enum KeyMode {
+  /** keyed with letters */
+  letters = 'letters', 
+  /** keyed with numbers */
+  numbers = 'numbers' 
+} 
+
+enum ChoiceMode {
+  checkbox, // user can select multiple choices
+  radio // user can  select one choice
+}
+
 /**
- * Config Object for @pie-elements/categorize
+* Model for the Choice Interaction
+*/
+export interface MultipleChoicePie {
+  
+  /** The choice options for the question */
+  choices: Choice[];
+  
+  /**  The question prompt or item stem*/
+  prompt?: string;
+  
+  /** What key should be displayed before choices. If undefined no  key will be displayed.  */
+  keyMode?: KeyMode; 
+  
+  /** Indicates the choices are single or multiple selection */
+  choiceMode?: ChoiceMode;
+  
+}
+
+
+/**
+ * Config Object for @pie-elements/multiple-choice
  */
-export interface CategorizePie  {
+export interface MultipleChoiceConfigure  {
 
   /** 
    * Whether config view will show a button that allows an author to add more choices
